+++ packages/beff-wasm/fixtures/codegen-snaps/export-decoder/bff-generated/parser.js
@@ -36,11 +36,6 @@ const buildParsersInput = {
 
 
 
-class BffParseError {
-  constructor(errors) {
-    this.errors = errors;
-  }
-}
 function buildParsers(args) {
 
   const customFormats = args?.customFormats ?? {}
@@ -53,6 +48,7 @@ function buildParsers(args) {
 
   Object.keys(customFormats).forEach((k) => {
     const v = customFormats[k];
+    
     registerCustomFormatter(k, v);
   });
 
@@ -62,8 +58,10 @@ function buildParsers(args) {
   Object.keys(buildParsersInput).forEach((k) => {
     
     let v = buildParsersInput[k];
-    const safeParse = (input) => {
-      const validatorCtx = {};
+    const safeParse = (input, options) => {
+      const validatorCtx = {
+        disallowExtraProperties: options?.disallowExtraProperties ?? false,
+      };
       const new_value = v(validatorCtx, input);
       const validation_result = validatorCtx.errors;
       if (validation_result == null) {
@@ -71,12 +69,12 @@ function buildParsers(args) {
       }
       return { success: false, errors: validation_result };
     };
-    const parse = (input) => {
-      const safe = safeParse(input);
+    const parse = (input, options) => {
+      const safe = safeParse(input, options);
       if (safe.success) {
         return safe.data;
       }
-      throw new BffParseError(safe.errors);
+      throw safe
     };
     const zod = () => {