+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @beff/client
 
+## 0.0.42
+
+### Patch Changes
+
+- disallow extra prop
+- Updated dependencies
+  - @beff/cli@0.0.42
+
+## 0.0.41
+
+### Patch Changes
+
+- ad-hoc b
+- Updated dependencies
+  - @beff/cli@0.0.41
+
+## 0.0.40
+
+### Patch Changes
+
+- move types to cli
+- Updated dependencies
+  - @beff/cli@0.0.40
+
 ## 0.0.39
 
 ### Patch Changes